@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:typed_data';
 
 import 'package:collection/collection.dart';
 import 'package:ed25519_edwards/ed25519_edwards.dart' as ed;
@@ -9,15 +10,15 @@ import 'package:test/test.dart';
 void main() {
   test('testSignVerify', () {
     var keyPair = ed.generateKey();
-    var privateKey = keyPair.privateKey;
-    var publicKey = keyPair.publicKey;
+    var privateKey = keyPair.privateKey!;
+    var publicKey = keyPair.publicKey!;
     var message = utf8.encode('test message');
-    var sig = ed.sign(privateKey, message);
+    var sig = ed.sign(privateKey, message as Uint8List);
     var result = ed.verify(publicKey, message, sig);
     assert(result == true);
 
     var wrongMessage = utf8.encode('wrong message');
-    var wrongResult = ed.verify(publicKey, wrongMessage, sig);
+    var wrongResult = ed.verify(publicKey, wrongMessage as Uint8List, sig);
     assert(wrongResult == false);
   });
 
@@ -32,36 +33,36 @@ void main() {
         .transform(utf8.decoder)
         .transform(LineSplitter())
         .forEach((l) {
-          index++;
-          print('line index: $index');
+      index++;
+      print('line index: $index');
 
-          var parts = l.split(':');
-          assert(parts.length == 5);
+      var parts = l.split(':');
+      assert(parts.length == 5);
 
-          var privBytes = HEX.decode(parts[0]);
-          var pubkey = HEX.decode(parts[1]);
-          var msg = HEX.decode(parts[2]);
-          var sig = HEX.decode(parts[3]);
-          sig = sig.sublist(0, ed.SignatureSize);
-          assert(pubkey.length == ed.PublicKeySize);
+      var privBytes = HEX.decode(parts[0]);
+      var pubkey = HEX.decode(parts[1]);
+      var msg = HEX.decode(parts[2]);
+      var sig = HEX.decode(parts[3]);
+      sig = sig.sublist(0, ed.SignatureSize);
+      assert(pubkey.length == ed.PublicKeySize);
 
-          var privkey = privBytes.sublist(0, ed.PrivateKeySize);
-          pubkey = privkey.sublist(32);
+      var privkey = privBytes.sublist(0, ed.PrivateKeySize);
+      pubkey = privkey.sublist(32);
 
-          var sig2 = ed.sign(ed.PrivateKey(privkey), msg);
-          assert(ListEquality().equals(sig, sig2));
+      var sig2 = ed.sign(ed.PrivateKey(privkey), msg as Uint8List);
+      assert(ListEquality().equals(sig, sig2));
 
-          var verified = ed.verify(ed.PublicKey(pubkey), msg, sig2);
-          assert(verified);
+      var verified = ed.verify(ed.PublicKey(pubkey), msg, sig2);
+      assert(verified);
 
-          var priv2 = ed.newKeyFromSeed(privkey.sublist(0, 32));
-          assert(ListEquality().equals(privkey, priv2.bytes));
+      var priv2 = ed.newKeyFromSeed(privkey.sublist(0, 32) as Uint8List);
+      assert(ListEquality().equals(privkey, priv2.bytes));
 
-          var pubkey2 = ed.public(priv2);
-          assert(ListEquality().equals(pubkey, pubkey2.bytes));
-          
-          var seed = ed.seed(priv2);
-          assert(ListEquality().equals(seed, privkey.sublist(0, 32)));
-        });
+      var pubkey2 = ed.public(priv2);
+      assert(ListEquality().equals(pubkey, pubkey2.bytes));
+
+      var seed = ed.seed(priv2);
+      assert(ListEquality().equals(seed, privkey.sublist(0, 32)));
+    });
   });
 }
